+++ ChangeLog.md
@@ -7,5 +7,14 @@
 | 0.03    | 24 APR 2025 | Scaremonger | Created Github |
 | 0.04    | 01 MAY 2025 | Scaremonger | fbtest02 finished |
 | 0.05    | 01 MAY 2025 | Scaremonger | Renamed tests |
-| 0.05.01 | 01 MAY 2025 | Scaremonger | Bugfix |
+| 0.05.01 | 01 MAY 2025 | Scaremonger | Bugfix | 
+
+| 0.06.00 | 01 MAY 2025 | Scaremonger | Created os.sigint library | 
+| 0.06.01 | 02 MAY 2025 | Scaremonger | Created linux.framebuffer library | 
+| 0.06.02 | 02 MAY 2025 | Scaremonger | Created alt.video library | 
+| 0.06.03 | 03 MAY 2025 | Scaremonger | Basic driver implemented and Example added to test it | 
+| 0.06.04 | 03 MAY 2025 | Scaremonger | Added plot() | 
+| 0.06.05 | 04 MAY 2025 | Scaremonger | Added backbuffer from initial proof-of-concept | 
+| 0.06.06 | 04 MAY 2025 | Scaremonger | Added putpixel(), Updated DrawHline() and UpdateVline() in TFramebuffer |
+| 0.06.07 | 05 MAY 2025 | Scaremonger | Fixed clear screen issue |